[K6.0] When uninstall Kunena template it doesn't found the contro method
