+++ src/admin/src/Controller/TemplatesController.php
@@ -200,73 +200,6 @@ class TemplatesController extends FormController
 	}
 
 	/**
-	 * Uninstall
-	 *
-	 * @return  void
-	 *
-	 * @throws  Exception
-	 * @throws  null
-	 * @since   Kunena 2.0
-	 */
-	public function uninstall(): void
-	{
-		$cid      = $this->app->input->get('cid', [], 'array');
-		$id       = array_shift($cid);
-		$template = $id;
-
-		if (!Session::checkToken())
-		{
-			$this->app->enqueueMessage(Text::_('COM_KUNENA_ERROR_TOKEN'), 'error');
-			$this->setRedirect(KunenaRoute::_($this->baseurl, false));
-
-			return;
-		}
-
-		// Initialize variables
-		$otemplate = KunenaTemplateHelper::parseXmlFile($id);
-
-		if (!$otemplate)
-		{
-			$this->app->enqueueMessage(Text::_('COM_KUNENA_A_TEMPLATE_MANAGER_TEMPLATE_NOT_SPECIFIED'), 'error');
-			$this->setRedirect(KunenaRoute::_($this->baseurl, false));
-
-			return;
-		}
-
-		if (\in_array($id, $this->locked))
-		{
-			$this->app->enqueueMessage(Text::sprintf('COM_KUNENA_A_CTRL_TEMPLATES_ERROR_UNINSTALL_SYSTEM_TEMPLATE', $otemplate->name), 'error');
-			$this->setRedirect(KunenaRoute::_($this->baseurl, false));
-
-			return;
-		}
-
-		if (KunenaTemplateHelper::isDefault($template))
-		{
-			$this->app->enqueueMessage(Text::sprintf('COM_KUNENA_A_CTRL_TEMPLATES_ERROR_UNINSTALL_DEFAULT_TEMPLATE', $otemplate->name), 'error');
-			$this->setRedirect(KunenaRoute::_($this->baseurl, false));
-
-			return;
-		}
-
-		$tpl = KPATH_SITE . '/template/' . $template;
-
-		// Delete the template directory
-		if (is_dir($tpl))
-		{
-			// Clear all cache, just in case.
-			KunenaCacheHelper::clearAll();
-			$this->app->enqueueMessage(Text::sprintf('COM_KUNENA_A_TEMPLATE_MANAGER_UNINSTALL_SUCCESS', $id));
-		}
-		else
-		{
-			$this->app->enqueueMessage(Text::_('COM_KUNENA_A_TEMPLATE_MANAGER_TEMPLATE') . ' ' . Text::_('COM_KUNENA_A_TEMPLATE_MANAGER_UNINSTALL') . ': ' . Text::_('COM_KUNENA_A_TEMPLATE_MANAGER_DIR_NOT_EXIST'));
-		}
-
-		$this->setRedirect(KunenaRoute::_($this->baseurl, false));
-	}
-
-	/**
 	 * Edit Scss
 	 *
 	 * @return  void
